\name{ENSO.dat}
\alias{ENSO.dat}
\title{Equatorial Pacific SST anomalies from El Niño 3 region.
}
\description{
The data set \code{\link{ENSO.dat}} contains an irregular time series (ENSO) 
with 125 data points and an average temporal spacing of 1.24 years covering 
the time interval 1850-2006. The ENSO data set come from Mann et al. (2009). 
The data sets can be obtained from the following URL \url{http://www.meteo.psu.edu/holocene/public_html/supplements/MultiproxySpatial09/results/}
(NINO3 full). 
}
\usage{data(ENSO)
}
\format{One file in ASCII format containing 125 elements and two variables (time and ENSO)
}
\source{\url{http://www.meteo.psu.edu/holocene/public_html/supplements/MultiproxySpatial09/results/} \cr

Mann, M. E., Zhang, Z., Rutherford, S., Bradley, R. S., Hughes, M. K.,
Shindell, D., Ammann, C., Faluvegi, G., Ni, F. (2009). Global signatures
and dynamical origins of the Little Ice Age and Medieval Climate Anomaly.
Science, \strong{326(5957)}, 1256-1260. <URL: \doi{10.1126/science.1177303}>. 
}
\keyword{dataset2}

