\name{ID31.dat}
\alias{ID31.dat}
\title{Unevenly-spaced pollen record from the marine sediments 
core (MD04-2845) collected on the southwestern European margin. 
}
\description{
The data set \code{\link{ID31.dat}} contains one paleoclimate 
(pollen percentages) time series spanning a time interval between 73,000 
and 15,000 years before present (BP), thus covering the last glacial period 
(LGP). This data set come from a global pollen and charcoal database 
(Sánchez Goñi et al., 2017) developed in the framework of the INQUA 
International Focus Group ACER (Abrupt Climate Changes and Environmental 
Responses). The paleoclimate time series come from the site MD04-2845 and contains 
77 elements (Sánchez Goñi et al., 2008, 2017).
}
\usage{data(MD04_2845_siteID31)
}
\format{One file in ASCII format containing 77 elements and two variables 
(time and pollen percentages). 
}
\source{\url{https://doi.pangaea.de/10.1594/PANGAEA.870867} \cr

Sánchez Goñi, M. F., Landais, A., Fletcher, W. J., Naughton, F., Desprat, S.,
Duprat, J. (2008). Contrasting impacts of Dansgaard-Oeschger events over
a western European latitudinal transect modulated by orbital parameters.
Quaternary Science Reviews 27 (11), 1136--1151. \cr

Sánchez Goñi, M. F., Desprat, S., Daniau, A.L., Bassinot, F. C., 
Polanco Martínez, J. M., Harrison, S. P., Allen, J. R., Anderson, R. S., Behling,
H., Bonnefille, R., et al. (2017). The ACER pollen and charcoal database: a
global resource to document vegetation and fire response to abrupt climate
changes during the last glacial period. Earth System Science Data 9 (2), 679. \cr
URL \url{https://www.earth-syst-sci-data.net/9/679/2017/}. 
}
\keyword{dataset2}

