\name{ID32.dat}
\alias{ID32.dat}
\title{Unevenly-spaced pollen record from the marine sediments 
core (MD95-2039) collected on the southwestern European margin. 
}
\description{
The data set \code{\link{ID32.dat}} contains a paleoclimate 
(pollen percentages) time series spanning a time interval between 73,000 
and 15,000 years before present (BP), thus covering the last glacial period 
(LGP). This data set come from a global pollen and charcoal database 
(Sánchez Goñi et al., 2017) developed in the framework of the INQUA 
International Focus Group ACER (Abrupt Climate Changes and Environmental 
Responses). The time series come from the site MD95-2039 and contains 141 
elements (Roucoux et al., 2005; Sánchez Goñi et al., 2017).
}
\usage{data(MD95_2039_siteID32)
}
\format{One file in ASCII format containing and 141 elements and two 
variables (time and pollen percentages). 
}
\source{\url{https://doi.pangaea.de/10.1594/PANGAEA.870867} \cr
Roucoux, K., De Abreu, L., Shackleton, N., Tzedakis, P. (2005). The response
of NW Iberian vegetation to North Atlantic climate oscillations during the
last 65 kyr. Quaternary Science Reviews 24 (14), 1637--1653. \cr

Sánchez Goñi, M. F., Desprat, S., Daniau, A.L., Bassinot, F. C., 
Polanco Martínez, J. M., Harrison, S. P., Allen, J. R., Anderson, R. S., Behling,
H., Bonnefille, R., et al. (2017). The ACER pollen and charcoal database: a
global resource to document vegetation and fire response to abrupt climate
changes during the last glacial period. Earth System Science Data 9 (2), 679.
}
\keyword{dataset2}

