\name{NHSST.dat}
\alias{NHSST.dat}
\title{Northern Hemisphere (NH) sea surface temperature (SST) anomalies.
}
\description{
The data set \code{\link{NHSST.dat}} contains an irregular time series (NH-SST) 
with 125 data points and an average temporal spacing of 1.24 years covering 
the time interval 1850-2006. The NH-SST data set come from HadCRUT3 (Brohan et 
al., 2006). The data sets can be obtained from the following URL \url{http://www.meteo.psu.edu/holocene/public_html/supplements/MultiproxySpatial09/results/}
(Northern Hemisphere full). 
}
\usage{data(NHSST)
}
\format{One file in ASCII format containing 125 elements and two variables (time and NHSST)
}
\source{\url{http://www.meteo.psu.edu/holocene/public_html/supplements/MultiproxySpatial09/results/} \cr

Brohan, P., Kennedy, J. J., Harris, I., Tett, S. F., Jones, P. D. (2006). 
Uncertainty estimates in regional and global observed temperature changes: 
A new data set from 1850. Journal of Geophysical Research: Atmospheres, 
\strong{111(D12)}.  <URL: \doi{https://doi.org/10.1029/2005JD006548}>.  
}
\keyword{dataset1}

